refactor(try): split code
