+++ packages/try/src/Painter/constants.ts
@@ -1,0 +1,3 @@
+export const Tool = ['pen', 'eraser'] as const;
+
+export type Tool = (typeof Tool)[number];
